resources fixed

+++ lib/dialogs/ingredients/custom_ing/custom_ing_form_dialog.dart
@@ -18,14 +18,14 @@ import 'package:provider/provider.dart';
 class CustomIngFormDialog extends StatefulWidget {
   final Ingredient? ingredient;
   final CustomIngredient? customIngredient;
-  final int quantity;
+  final double quantity;
   final Unit? unit;
   final List<Category> categories;
   final Function(
     String name,
     Category category,
     List<Tag> tags,
-    int quantity,
+    double quantity,
     Unit unit,
   )
   onSave;
@@ -299,7 +299,7 @@ class _CustomIngFormDialogState extends State<CustomIngFormDialog> {
                     _nameController.text,
                     _selectedCategory,
                     _selectedTags.toList(),
-                    int.parse(_quantityController.text),
+                    double.parse(_quantityController.text),
                     _selectedUnit,
                   );
                   if (mounted) Navigator.pop(context);
